V0.6.1.3m
Refactor:
- TooltipAutoSetup split into multiple classes
- TooltipSystem separated
- TooltipTrigger moved to отдельный скрипт
- TooltipFactory added for UI creation

Improve:
- Code structure optimized (OOP)
- Responsibilities separated
